requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via realtime
aiosignal==1.3.2
    # via aiohttp
alembic==1.14.1
    # via optuna
amqp==5.3.1
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   starlette
attrs==25.1.0
    # via aiohttp
billiard==4.2.1
    # via celery
boto3==1.36.12
    # via -r requirements.in
botocore==1.36.12
    # via
    #   boto3
    #   s3transfer
celery==5.4.0
    # via -r requirements.in
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via
    #   click
    #   colorlog
    #   tqdm
colorlog==6.9.0
    # via optuna
deprecation==2.1.0
    # via postgrest
faiss-cpu==1.10.0
    # via -r requirements.in
fastapi==0.115.8
    # via -r requirements.in
filelock==3.17.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.2.0
    # via
    #   huggingface-hub
    #   torch
gotrue==2.11.3
    # via supabase
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
h2==4.2.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.7
    # via httpx
httpx[http2]==0.28.1
    # via
    #   gotrue
    #   postgrest
    #   storage3
    #   supabase
    #   supafunc
huggingface-hub==0.28.1
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.5
    # via torch
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joblib==1.4.2
    # via scikit-learn
kombu==5.4.2
    # via celery
mako==1.3.8
    # via alembic
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mpmath==1.3.0
    # via sympy
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
networkx==3.4.2
    # via torch
numpy==2.2.2
    # via
    #   -r requirements.in
    #   faiss-cpu
    #   optuna
    #   pandas
    #   rank-bm25
    #   scikit-learn
    #   scipy
    #   torchvision
    #   transformers
optuna==4.2.0
    # via -r requirements.in
packaging==24.2
    # via
    #   deprecation
    #   faiss-cpu
    #   huggingface-hub
    #   optuna
    #   transformers
pandas==2.2.3
    # via -r requirements.in
pillow==11.1.0
    # via
    #   sentence-transformers
    #   torchvision
postgrest==0.19.3
    # via supabase
prompt-toolkit==3.0.50
    # via click-repl
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
psycopg2==2.9.10
    # via -r requirements.in
pydantic==2.10.6
    # via
    #   fastapi
    #   gotrue
    #   postgrest
pydantic-core==2.27.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   celery
    #   pandas
    #   realtime
    #   storage3
pytz==2025.1
    # via pandas
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   optuna
    #   transformers
rank-bm25==0.2.2
    # via -r requirements.in
realtime==2.3.0
    # via supabase
redis==5.2.1
    # via -r requirements.in
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   transformers
s3transfer==0.11.2
    # via boto3
safetensors==0.5.2
    # via transformers
scikit-learn==1.6.1
    # via sentence-transformers
scipy==1.15.1
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==3.4.1
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.37
    # via
    #   alembic
    #   optuna
starlette==0.45.3
    # via fastapi
storage3==0.11.3
    # via supabase
strenum==0.4.15
    # via supafunc
supabase==2.12.0
    # via -r requirements.in
supafunc==0.9.3
    # via supabase
sympy==1.13.1
    # via torch
threadpoolctl==3.5.0
    # via scikit-learn
tokenizers==0.21.0
    # via transformers
torch==2.6.0
    # via
    #   -r requirements.in
    #   sentence-transformers
    #   torchvision
torchvision==0.21.0
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   optuna
    #   sentence-transformers
    #   transformers
transformers==4.48.2
    # via sentence-transformers
typing-extensions==4.12.2
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   realtime
    #   sqlalchemy
    #   torch
tzdata==2025.1
    # via
    #   celery
    #   kombu
    #   pandas
urllib3==2.3.0
    # via
    #   botocore
    #   requests
uvicorn==0.34.0
    # via -r requirements.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
websockets==14.2
    # via realtime
yarl==1.18.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
